GoogleEarthProImagePipeline/requirements.txt
pandas
numpy
skimmage
simplekml
shapely
osgep
tifffile 
qgis.core
pyproj
geopandas